# api/api-actor.swagger.yaml
openapi: 3.0.0
info:
    title: 'Sakila API Documentation'
    version: 1.0.0
    description: 'API documentation for Film rental store management system'
servers:
    - {url: 'http://localhost:3000', description: 'Development server'}
paths:
    /v1/actors: {get: {summary: 'Returns a list of all actors', tags: [Actors], responses: {'200': {description: 'List of actors', content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Actor'}}}}}}}, post: {summary: 'Create a new actor', tags: [Actors], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ActorInput'}}}}, responses: {'201': {description: 'Actor created successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Actor'}}}}, '400': {description: 'Invalid input', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}}
    '/v1/actors/{id}': {get: {summary: 'Get actor by ID', tags: [Actors], parameters: [{in: path, name: id, schema: {type: string}, required: true, description: 'Actor ID'}], responses: {'200': {description: 'Actor found', content: {application/json: {schema: {$ref: '#/components/schemas/Actor'}}}}, '404': {description: 'Actor not found', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}, delete: {summary: 'Delete an actor', tags: [Actors], parameters: [{in: path, name: id, schema: {type: string}, required: true, description: 'Actor ID'}], responses: {'200': {description: 'Actor deleted successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Actor'}}}}, '404': {description: 'Actor not found to be deleted', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}, put: {summary: 'Update an actor', tags: [Actors], parameters: [{in: path, name: id, schema: {type: string}, required: true, description: 'Actor ID'}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ActorInput'}}}}, responses: {'200': {description: 'Actor updated successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Actor'}}}}, '400': {description: 'Invalid input', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '404': {description: 'Actor not found to be updated', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}}
    /v1/films: {get: {summary: 'Get all films', tags: [Films], responses: {'200': {description: 'Retrieve all films successfully', content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Film'}}}}}, '500': {description: 'Internal Server Error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}, post: {summary: 'Create a new film', tags: [Films], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, responses: {'201': {description: 'Film created successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, '400': {description: 'Invalid input', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '500': {description: 'Internal Server Error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}}
    '/v1/films/{id}': {get: {summary: 'Get film by ID', tags: [Films], parameters: [{in: path, name: id, required: true, schema: {type: integer, minimum: 1}, description: 'Film ID'}], responses: {'200': {description: 'Retrieve film successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, '404': {description: 'Film not found', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '500': {description: 'Internal Server Error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}, delete: {summary: 'Delete film by ID', tags: [Films], parameters: [{in: path, name: id, required: true, schema: {type: integer, minimum: 1}, description: 'Film ID'}], responses: {'200': {description: 'Film deleted successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, '404': {description: 'Film not found', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '500': {description: 'Internal Server Error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}, put: {summary: 'Update film by ID', tags: [Films], parameters: [{in: path, name: id, required: true, schema: {type: integer, minimum: 1}, description: 'Film ID'}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, responses: {'200': {description: 'Film updated successfully', content: {application/json: {schema: {$ref: '#/components/schemas/Film'}}}}, '400': {description: 'Invalid input', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '500': {description: 'Internal Server Error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}}}
components:
    schemas: {Actor: {type: object, required: [firstName, lastName], properties: {actorId: {type: integer, minimum: 1, readOnly: true, description: 'Auto-generated ID of the actor'}, firstName: {type: string, description: 'First name of the actor'}, lastName: {type: string, description: 'Last name of the actor'}, lastUpdate: {type: string, format: date-time, readOnly: true}}}, ActorInput: {type: object, required: [firstName, lastName], properties: {firstName: {type: string, description: 'First name of the actor'}, lastName: {type: string, description: 'Last name of the actor'}}}, Error: {type: object, required: [errorId, message], properties: {errorId: {type: string, description: 'Error code of errors'}, message: {type: string, description: 'Error message of errors'}}}, Film: {type: object, required: [title, languageId], properties: {filmId: {type: integer, readOnly: true, minimum: 1}, title: {type: string, minLength: 1, maxLength: 255}, description: {type: string, nullable: true}, releaseYear: {type: integer, minimum: 1900, maximum: 2099, example: 2025, nullable: true}, languageId: {type: integer, minimum: 1}, originalLanguageId: {type: integer, minimum: 1, nullable: true}, rentalDuration: {type: integer, minimum: 1, default: 3}, rentalRate: {type: number, format: float, minimum: 0, default: 4.99}, length: {type: integer, minimum: 1, nullable: true}, replacementCost: {type: number, format: float, minimum: 0, default: 19.99}, rating: {type: string, enum: [G, PG, PG-13, R, NC-17], nullable: true}, specialFeatures: {type: string, nullable: true, example: null}, lastUpdate: {type: string, format: date-time, readOnly: true}}}}
tags: []